Fixed issue with blocks popup menu displayed only once with Lazarus 1.8.
The block popup menu is triggered automatically by the LCL widget set
code. Triggering it inside the on mouse down event causes the new LCL
to enter an endless loop. This behavior was hidden in previous LCL
versions by a bug allowing multiple popup menus to be displayed at the
same time.